Fix logic for filtering out variants
Since the personalisable page metadata doesn't get populated for newly
created pages (that aren't variants) until it's first accessed, newly
created pages don't appear in the page list, because they satisfy the
filter rule with None==None.

One approach to fix this would be a post-save hook to initialise the
metadata for newly created pages, but fixing the filtering logic, as
done here is probably simpler. One the pages appear in the list, their
metadata gets initialised anyway, and users are taken to the list after
they create a page.